moved shuffle func to utils
