+++ src/pages/Projects.tsx
@@ -3,19 +3,11 @@ import { MdCancel } from "react-icons/md";
 import ProjectCard from "../components/ProjectCard";
 import { makeRequest } from "../util/backend";
 import { IEndpointTypes } from "../util/types";
+import { shuffle } from "../util/shuffle";
 import Fuse from "fuse.js";
 import SpinnerLoader from "../components/SpinnerLoader";
 import { IconContext } from "react-icons";
 
-// function to randomly shuffle an array
-function shuffle(array) {
-  for (var i = array.length - 1; i > 0; i--) {
-      var j = Math.floor(Math.random() * (i + 1));
-      var temp = array[i];
-      array[i] = array[j];
-      array[j] = temp;
-  }
-}
 
 function Projects() {
   const [projects, setProjects] = useState<
@@ -40,9 +32,9 @@ function Projects() {
 
   useEffect(() => {
     makeRequest("project", "get")
-      .then(async (response) => {
+      .then((response) => {
         if (response.is_ok) {
-          await shuffle(response.response);// shuffles the array
+          shuffle(response.response);// shuffles the array
           setProjects(response.response);
         } else {
           setError("Error fetching projects.");
